+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/perks/Perks.java
@@ -1,7 +1,6 @@
 package com.robertx22.age_of_exile.aoe_data.database.perks;
 
 import com.robertx22.age_of_exile.database.OptScaleExactStat;
-import com.robertx22.age_of_exile.database.data.perks.Perk;
 import com.robertx22.age_of_exile.database.data.stats.types.core_stats.Dexterity;
 import com.robertx22.age_of_exile.database.data.stats.types.core_stats.Intelligence;
 import com.robertx22.age_of_exile.database.data.stats.types.core_stats.Strength;
@@ -35,12 +34,6 @@ public class Perks implements ISlashRegistryInit {
 
     @Override
     public void registerAll() {
-
-        Perk start = new Perk();
-        start.is_entry = true;
-        start.identifier = "start";
-        start.type = Perk.PerkType.START;
-        start.addToSerializables();
 
         PerkBuilder.stat("int", new OptScaleExactStat(10, Intelligence.INSTANCE, ModType.FLAT));
         PerkBuilder.stat("dex", new OptScaleExactStat(10, Dexterity.INSTANCE, ModType.FLAT));