+++ src/main/java/com/robertx22/age_of_exile/saveclasses/unit/Unit.java
@@ -457,7 +457,7 @@ public class Unit {
 
                 if (entity instanceof PlayerEntity) {
 
-                    if (!gear.meetsStatRequirements(data)) {
+                    if (!gear.canPlayerWear(data)) {
                         addstats = false;
                     }